Move get_countrytag test
