+++ cartoframes/client/__init__.py
@@ -1,9 +1,10 @@
 from __future__ import absolute_import
 
 from .sql_client import SQLClient
-from .data_obs_client import DataObsClient
+from .data_obs_client import DataObsClient, get_countrytag
 
 __all__ = [
     'SQLClient',
-    'DataObsClient'
+    'DataObsClient',
+    'get_countrytag'
 ]
